requirements.txt
gradio>=4.0.0
google-cloud-vision>=3.5.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0